+++ CHANGELOG.md
@@ -7,7 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ## [Unreleased]
 
-Nothing at the moment
+### Fixed
+
+- Add support for buyer addresses with 2 lines (#2)
 
 
 ## [0.1.0] - 2019-08-31